refactor(db) split iteration logic into a separate module
Splits the `row_iterator` iteration logic used by `:each`
and `:each_for_*` into a separate module, so it can be
used in strategy level as well. This is a preparation for
implementing cascade-delete for Cassandra.
