+++ kong/db/dao/init.lua
@@ -1,4 +1,5 @@
-local cjson     = require "cjson"
+local cjson = require "cjson"
+local iteration = require "kong.db.iteration"
 
 
 local setmetatable = setmetatable
@@ -150,85 +151,6 @@ local function validate_options_value(options, schema, context)
 end
 
 
-local function iteration_failed(err, err_t)
-  local failed = false
-  return function()
-    if failed then
-      return nil
-    end
-    failed = true
-    return false, err, err_t
-  end
-end
-
-
-local function page_iterator(pager, size, options)
-  local page = 1
-  local i, rows, err, offset = 0, pager(size, nil, options)
-
-  return function()
-    if not rows then
-      return nil, err
-    end
-
-    i = i + 1
-
-    local row = rows[i]
-    if row then
-      return row, nil, page
-    end
-
-    if i > size and offset then
-      i, rows, err, offset = 1, pager(size, offset, options)
-      if not rows then
-        return nil, err
-      end
-
-      page = page + 1
-
-      return rows[i], nil, page
-    end
-
-    return nil
-  end
-end
-
-
-local function row_iterator(self, pager, size, options)
-  local next_row = page_iterator(pager, size, options)
-
-  local failed = false -- avoid infinite loop if error is not caught
-  return function()
-    local err_t
-    if failed then
-      return nil
-    end
-    local row, err, page = next_row()
-    if not row then
-      if err then
-        failed = true
-        if type(err) == "table" then
-          return false, tostring(err), err
-        end
-
-        err_t = self.errors:database_error(err)
-        return false, tostring(err_t), err_t
-      end
-
-      return nil
-    end
-
-    row, err, err_t = self:row_to_entity(row)
-    if not row then
-      failed = true
-      return false, err, err_t
-    end
-
-    return row, nil, page
-  end
-end
-
-
 local function check_update(self, key, entity, options, name)
   local entity_to_update, err, read_before_write =
     self.schema:process_auto_fields(entity, "update")
@@ -351,7 +273,7 @@ local function generate_foreign_key_methods(schema)
           local ok, err = validate_size_value(size)
           if not ok then
             local err_t = self.errors:invalid_size(err)
-            return iteration_failed(tostring(err_t), err_t)
+            return iteration.failed(tostring(err_t), err_t)
           end
 
         else
@@ -361,7 +283,7 @@ local function generate_foreign_key_methods(schema)
         local ok, errors = self.schema:validate_primary_key(foreign_key)
         if not ok then
           local err_t = self.errors:invalid_primary_key(errors)
-          return iteration_failed(tostring(err_t), err_t)
+          return iteration.failed(tostring(err_t), err_t)
         end
 
         local strategy = self.strategy
@@ -369,7 +291,7 @@ local function generate_foreign_key_methods(schema)
         local pager = function(size, offset)
           return strategy[page_method_name](strategy, foreign_key, size, offset)
         end
-        return row_iterator(self, pager, size)
+        return iteration.by_row(self, pager, size)
       end
 
     elseif field.unique or schema.endpoint_key == name then
@@ -701,7 +623,7 @@ function DAO:each(size, options)
   local pager = function(size, offset, options)
     return self.strategy:page(size, offset, options)
   end
-  return row_iterator(self, pager, size, options)
+  return iteration.by_row(self, pager, size, options)
 end
